Add ability to inject/remove extra repos
This patch adds to two new variables to Ansible, and Packer support for
them when building OVAs, AMI, and on GCE.

extra_repos: This is a comma delimited list of files to copy into the
build target as repo definitions. For CentOS, these should be .repo
files and will go to /etc/yum.repos.d, and for Ubuntu these should be
.list files and will go to /etc/apt/sources.list.d.

The list of files can use relative or absolute paths. If using relative
paths, the paths must reference files within Ansible's search
directories.

remove_extra_repos: This is a boolean flag that controls whether the
repo files added via `extra_repos` will be deleted during the sysprep
task. The default is `false`.