OVA: Make CentOS kickstart not rely on internet
This patch changes the ova centos kickstart to only install packages
from the cdrom/iso. We are already relying on Ansible as a follow-on
provisioner to install required packages after the OS is installed, so
these steps aren't really necessary. There was also a good bit of
overlap where the Ansible tasks were installing packages that were
listed in the kickstart -- this just leaves it to Ansible.

This makes building in enterprise environments w/o internet access
simpler as well.