OVA: parameterize network name
A recent commit hardcoded the network name to "VM Network". This
appears to work fine when building with Fusion, but if building
Workstation and that network is not found, there is no network connected
and the VM fails to boot.

Rather than hardcoding the `vmx_data` field, we can just set `network`
instead, and go back to the default value of "". Then this can be
overridden if a user needs to use a specific network by name.