Fix CentOS ISO download link
The new link put in place the previous day was incorrect. Despite having
tested locally, it appears that Packer looked to see if a file with the
correct checksum was in place (it was), and therefore dididn't notice
the link was wrong. It wasn't until running on a machine w/o the ISO
already downloaded the mistake was noticed.