List Photon RPMs as a list instead of a string
The list of Photon RPMs is much easier to read as a list rather than one
long string. It also follows the same pattern as the CentOS PRMs and
Debian packages. It was done as a string since it's a raw call to `tdnf`
instead of having a tdnf Ansible module, but we can instead just
concatenate the list into a string beforehand.