Ubuntu: Disable apt-daily and unattended-upgrades
By default, an apt-daily, apt-daily-upgrade, and unattended-upgrades
service was running on the Ubuntu images. The apt-daily services ran on
a timer, and automatically do something akin to "apt-get update" so that
users may not have to worry about it. Since these images are immutable,
there is no need for this step, and in fact it can contribute to minutes
of additional boot-up time.

Additionally, disabling unattended upgrades is desired for immutability,
as the system should not be updating any packages automatically.