OVA: Disable Hyper-V KVP protocol daemon
Ubuntu 18.04 has a bug that causes the startup of the Hyper-V KVP
protocol Daemon to fail during boot. This service startup as a 90 second
timeout, and the boot process does not continue until this timeout
expires. There is no need for OVAs that are generated to run on ESXi to
run a daemon specific to Hyper-V, so we can disable it entirely. This
regains 90 seconds of boot time.