No longer call post-create script during OVA build
When building OVA images, after the OVA is packaged up, a call to
images/capi/hack/image-post-create-config.sh is made. This convenience
script modifies the VM size and sets up with some useful cloud-init data
for developer debugging.

However, this assumes that the VM (VMX file) that it operates on is
loaded/imported to your VM managment software. If you are using VMW
Workstation, this may work well for you. If you are using VMW Fusion, it
does not. The VM created does not show up automatically in Fusion. I end
up having to import teh disk to FUsion, and then run this script
manually to set it up.

My point is that the Packer config is optimized for the 5% case, I
think. I pretty much never use this setup. When I do, I end up having to
run it manually anyways, so I think it is a wasted step.