Move common ansible args to one location
Right now, each CAPI image builder (AWS, Azure, DigitalOcean, GCE,
vSphere) has a separate definition of the "extra-vars" to send to
Ansible. These are in fact mostly identical, and they should be.
Therefore, move the defintion to a common location.

The vSphere builder has a couple of unique extra vars that the other
builders don't need, so the ability to have provider-specific
additions is preserved (and used for vSphere).