Use a single user/pass across multiple distros
When building the OVA images, a user and password is establed as part of
the kickstart/preseed. This user is then used for Ansible to provision
the node, and that user is locked out at the very end. It's never used
again. As we expand the number of supported distros, it seems like
extraneous effort to set a different user/pass for each distro. It also
makes re-use of kickstart files produce odd circumstances, such as
creating a "centos" user on a rhel machine, for example.

This patch standardizes the user to a non-distro specific one, and
instead just uses builder/builder.