Add ability to inject extra packages
This patch exposes the ability for a user to provide a list of extra RPM
or DEB files to install during the build.

We also document the "proper" way to do this -- by providing a JSON file
to the build rather than editing existing JSON files or the Ansible
YAML files.