Place user supplied var files at end of command
This patch makes sure that any user-supplied files that will be listed
out as a -var-file for Packer will be listed at the very end of the list
of var files. This is because if a var is defined more than once, the
last supplied one takes precedence. We want users to be able to override
defaults, and that cannot happen without this change.

We also make sure that the common config files found in packer/config
are also listed first.