ova: remove capi version
The OVA is the only image that trieds to embed some notion of the CAPI
version to the image. However, there is nothing in the image related to
CAPI, and nothing that changes when a new version of CAPI is released.
Embedding this version in the image is misleading at best, and can
easily go stale.