ova: correct XML syntax for OVF
While the current syntax is valid XML, reading in the XML with an XML
paring library (such as lxml from Python) and writing the exact same
thing back out to file produces a diff. The current text is inconsistent
with standard libraries, and also differs from what will be generated
from VMware's OVFtool. This patch fixes that syntax.