ova: replace OVF metadata for ESX builds
When using the esx build targets, the image is exported from ESX as an
OVA. Because of this, the tooling in place to create an OVA from scratch
was skipped, meaning that none of the metadata that was set by that
tooling was present.

This patch takes the exported OVA, extracts it, and puts the same
OVF definition in place that's used when building on a local desktop
hypervisor.