Force cri-tools symlink
When Kubernetes is not installed via package, there is an Ansible
snipped that symlinks cri-tools from /usr/local/bin to /usr/bin. With
the latest Photon packages installed, a newer version of Docker is
pulled in, which now requires containerd. This puts a binary at
/usr/bin/ctr, and when we try installer the *newer* containerd and
symlink from /usr/local/bin, Ansible refuses to overwrite it. This patch
forces the overwrite.