Add haproxy and pki roles
For the HAProxy OVA, add roles for installing and configuring HAProxy,
cloud-init, and the needed PKI. The HAProxy role makes re-use of the
setup role as a role dependency to install RPMs, work with extra repos,
etc. Cloud-init is installed by using the providers role, which for
VMware, already does exactly what we need.