ova: no need to symlink cloud-init in CentOS
With the latest CentOS 7 builds, the step to symlink cloud-init was
failing because the symlink target already existed. It already existed
because it is put there by the cloud-init package itself. I can only
guess this was a packaging change at some point, but this step is no
longer needed.