Bump OVA Ubuntu 1804 to 18.04.4
Bump the Ubuntu ISO to 18.04.4. As part of this change, it appears that
unattended upgrades is no longer installed by default. I've been unable
to determine what was pulling it in before, and what the difference is.
Since it is no longer installed by default, hedge against it being
present in older versions and future versions by making the disablement
of it conditional on it's presence as a separate step in the sysprep
role.