capi: Add ability to run user-provided role
This patch adds the ability for a user to specify their own custom
Ansible role that will be run right before the "sysprep" role. This
would allow for free-form customization of the image that is not
possible with the currently exposed variables/extension points. Users
can run their own roles without modifying the existing ones, and avoid
creating a dirty git repo at build time.

Additionally, the use of the custom role is captured in OVA metadata,
such that it can be determined upon inspection whether any additional
steps were possible. Use of the custom role means that more or less
anything could be done to the image, so it's something that should be
advertised and may indicate lack of supportability. We need to start
capturing similar metadata for additional image formats (e.g. AMIs).