Bump Photon ISO to Update2 Revision 2
The hosting location has also moved from Bintry to packages.vmware.com.

Due to a quirk in how Packer checks for whether or not the ISO is
already in the Packer cache (it uses the ISO filename in the hash
creation, not the URL, which makes *total* sense since your or normally
pointing to the same ISO located on different mirrors), Packer thinks
this ISO is already in the cache if you had previously downloaded
Update2 Revision 1. Why? Because it's the same filename. I have no idea
why it is, but it is.

So if, like me, you get really confused why it didn't have to download a
new ISO when first switching to this, look for
c9ab401e209feb6e1ff8466179b4569c33c1f9b2.iso in your packer cache. if
it's there, remove it! Packer will *not* download the new ISO and verify
its checksum if you do not.