Fix default EPEL location for CentOS 7.
A recent change made it possible to set the EPEL location to an empty
string to skip installing EPEL. Since this var is now always set by
Packer, not setting it makes it be an empty string, and since extra-vars
take precedence in Ansible, the net effect is that the var was *always*
set to an empty string, even though a default is set in the Ansible
role.

This patch sets the default value for CentOS 7 back as it was. The
addition is only the CentOS 7 file for OVAs, so it does not effect AL2,
and if needed in a future CentOS 8 distro, a release specific version
can be set there as well.