Disable Hyper-V KVP daemon for Ubuntu AMI
When launching Ubuntu 18.04 AMIs, it can be seen from the console
output that that a kernel daemon fails on startup, with the name
"Dependency failed for Hyper-V KVP Protocol Daemon." This adds over 60
seconds of boot time to the instance. Since AMIs are only running on
AWS, there is no need for a Hyper-V daemon, and it can safely be
disabled. The same steps were taken for OVAs/ESX images months ago.