Fix ensure Py3, Pip3, and azure-cli
This patch de-dedups logic for making sure that Python3 and pip3 are
installed. It also corrects a corner case where pip3 could be installed
to the user local Python bin dir but that dir was not in path.

And finally, for checking that azure-cli was installed (which goes to
the user local Python bin dir), the wrong command name was used. It was
looking for azure-cli when it should look for az.