photon: Break out reboot into separate shell provisioner step.
When using Packer, Ansible roles should not include the reboot command.
This is because Packer controls the SSH connection, and has no indication
that the machine will be rebooted - the SSH connection is reset and no
attempt to reconnect is made. This can be obviated with some hacks, but
here instead, we create a specific firstboot playbook, and insert a
shell provisioner step to reboot the machine using the "expect_disconnect"
flag to inform Packer that the machine will disconnect and it can move
on to the next provisioner.

Signed-off-by: Naadir Jeewa <jeewan@vmware.com>