capi: fail if containerd version does not match
In case of Flatcar, if the given containerd version does not match with
the running containerd version, the image-builder should fail hard.
A version mismatch with the current version.

So it means we need to always set a correct version from Packer side, to
make the containerd installation work for Flatcar.