capi: sort Flatcar versions correctly
As soon as Flatcar stable
[2605.10.0](https://www.flatcar-linux.org/releases/#release-2605.10.0)
was released, the script `image-grok-latest-flatcar-version.sh` stopped
working as expected. What happens is, the json
[feed](https://www.flatcar-linux.org/releases-json/releases-stable.json)
returns a list of versions with `2605.9.0` at the end, not `2605.10.0`,
because the list is sorted alphanumerically. To convert the list into a
semver-sorted one, we should apply a filter `sort --version-sort`, not
a pure `sort`.