Only place required files in Docker container
This patch changes up the Dockerfile and .dockerignore to start with an
explicity deny all approach, and then add only the files/dirs that are
needed. This is most useful for developers, who may have additional
files/folders located withiin images/capi that are not part of the git
repo. Since the old Docker file was doing "COPY .", this copied anything
that wasn't explicitly ruled out in .dockerignore. This patch takes the
opposite approach.