Use OVFTool to create OVAs
This patch is a major change for users who are building OVAs. With
recent versions of vSphere and Fusion (and likely Workstation, but that
was untested), the OVAs created by image-builder were no longer
importable from local disk (but oddly the worked if referenced as a
URL). The root cause of this remains a bit of a mystery -- the problem
lies with the OVA file, not the OVF. A problematic OVA could be
untarred, then the OVF used as a source for import and the import was
successful.

It has long been discussed and disired that ovftool should be used to
create the OVA instead of using `tarfile` from Python. These recent
errors are reason enough to make the switch. Using ovftool has the
advantage of validating the OVF file itself as well, such that if there
are any issues with the file then the OVA will not get created.

Installing ovftool requires accepting a EULA and must be installed by a
user on their system. The `deps-ova` target has been updated to reflect
this, and directs users where to download from.