ova: don't use K8s version in base naming
When building base images for use with OVAs, the base image is solely
based on the OS install ISO and does not include anything for
Kubernetes. The only time the base image needs to change is if the ISO
changes or the kickstart/preseed changes. We were previously naming the
base image with the K8s version included. This change removes that,
such that the base names are only as "base-<OS>".