Fix validate make targets
There were a few problems with the validate make targets. Flatcar had
been added to the list of OS's supported for OVA builds, which is
incorrect. The Azure windows-sig target had a typo in it. The QEMU
target was missing flatcar, and the AMI target wasn't documented. This
patch fixes all those things and `make validate-all` can work again,
presubming you supply the needed ENV vars.

A CI script is added that automatically adds the required ENV vars and
can also be used by Prow/CI.