Clean up and *use* the ansible.cfg file
The ansible.cfg file that is present in images/capi/ansible is not used
in any way because the Ansible commands won't pick it up in that
location. Since Ansible is run with a PWD of images/capi, put it there
instead.

Additionally, the .cfg file mistakenly had 'default' instead of
'defaults', meaning that all entries were being ignored. We no longer
need to specify a search path for filter plugins since we haven't used
any in a while, and the setting for retry files is already the default
in Ansible.