Add CI script for populating GOSS tests
The GOSS tests that run are generated dynamically from a combination of
a GOSS vars file, and inline vars from the Packer files. Since the GOSS
files have loops in them that look for specific keys, it is possible
that missing keys cause a failure in test generation. This CI test
ensures that the GOSS vars file has all required keys and that the
templates render successfully.

Right now, Flatcar does not use the same mechanism as all other OS's,
but this CI job tests it as if it does, so that that transition is
possible if we ever try that.