capi: remove machine ID for systemd first boot semantics on Flatcar
Ignition enables systemd services through systemd preset files. These
preset files are only evaluated when systemd considers the machine to
have a first boot. While the machine ID was already reset by truncating
the file, this is not enough to make systemd also evaluate the presets
again, which only happens when the file is removed. Without this, the
systemd services that Ignition enables end up disabled.

Remove the machine ID on Flatcar to ensure the Ignition systemd preset
gets evaluated. For the presets that Flatcar ships on the /usr
partition there is no harm to evaluate them again because pre-enabled
services on Flatcar can only be disabled by masking, and if Ansible
masked a service this will be kept even if a preset would want to
enable the service. Since regular service disabling works on other
distributions there is a potential harm of reevaluationg the presets
because they could enable the service again, thus, only remove the
machine ID file for Flatcar.