capi: disable OS update for Flatcar Container Linux
The image is built from a specific Flatcar release but when
update-engine is left enabled, the instance will update to the latest
Flatcar release as soon as it is available. This may cause unexpected
issues and introduces disruptions because a newly deployed node will
reboot immediately after the first boot when an update is available and
later on each time a new update is published.

Disable update-engine and locksmithd by default to freeze the Flatcar
Container Linux version and prevent issues coming from unscheduled
updates. Users can still enable updates again if they want to and set
up an update coordinator such as FLUO.