capi: fully install containerd in Flatcar instead of using inbuilt
The mixture of using the inbuilt containerd and relying on specific
aspects on how containerd present through Torcx on Flatcar, combined
with requiring a particular containerd version does not work well as
soon as Flatcar Container Linux ships containerd differently.

To align more with what the other distributions do, download and
install containerd to /opt/ in Flatcar (/usr/ is read-only), and use
the default systemd service file except for the corrected paths
which need to be set through a service drop-in file. This also has the
effect that the default containerd socket location is used and for
Docker on Flatcar we have to set up a symlink.
Since the inbuilt containerd binaries are still present under /usr/,
manual shell usage needs to ensure that the PATH is set to prefer
/opt/bin/ (export PATH="/opt/bin:$PATH").