capi: set cgroup v1 in Flatcar to match containerd/kubelet config
The current Stable releases of Flatcar container Linux are defaulting
to v1 but soon an update to using v2 by default will come. The
containerd/kubelet cgroup driver is chosen based on the Kubernetes
version being >= 1.21. The cgroup driver must be "systemd" for v2,
while for v1 both are possible - in case "systemd" is not chosen,
we have to ensure that the OS runs in cgroup v1.

Force running in cgroupv1 if the "systemd" driver is not chosen (K8s
version < 1.21.0) and otherwise leave the OS default.