capi: ensure first_boot marker is created for Flatcar
The Azure packer script for Flatcar uses a different way to bootstrap python
onto the instance, which does not go through the existing bootstrap-flatcar.sh
script. This exposed the fact that bootstrap-flatcar.sh is the only place where
/boot/flatcar/first_boot is created. This marker file is used by Flatcar to
ensure the provisioning application (ignition) runs on boot, and prevented
produced Azure images from interpreting customData when provisioned again.

Ensure the marker file gets created in the sysprep phase so that the image can
be provisioned again.