capi: azure: execute 'waagent -deprovision' for Flatcar
Flatcar uses ignition+waagent for provisioning. Waagent is responsible for
configuring the instance hostname and admin user ssh keys. Waagent only
performs provisioning if /var/lib/waagent/provisioned doesn't exists. Therefore
we must ensure that the instance is deprovisioned during building so that it
will be reprovisoned when used as an image for new instances. 'waagent -deprovision+user'
ensures that the image is generalized and the build user (packer) is removed,
and will allow a new hostname and admin user ssh-key to be set when next
provisioned.