capi: ensure 127.0.0.1 is first in /etc/hosts
When Flatcar Container Linux is used with upstream containerd binaries, when a
lookup for localhost is performed only the first match in /etc/hosts is
returned. This results in portforwarding to etcd failing with the following
messages:

  containerd[1003]: time="2021-09-22T13:09:39.189644284Z" level=info msg="Executing port forwarding in network namespace \"host\""
  containerd[1003]: E0922 13:09:39.190414    1003 httpstream.go:257] error forwarding port 2379 to pod b0b39725d5e18ff487dd4f9e9f04d6f84dd33e6c13d2a7a77cbb33694efc924a, uid : failed to execute portforward in network namespace "host": failed to connect to localhost:2379 inside namespace "b0b39725d5e18ff487dd4f9e9f04d6f84dd33e6c13d2a7a77cbb33694efc924a", IPv4: dial tcp4: lookup localhost: No address associated with hostname IPv6 dial tcp6 [::1]:2379: connect: connection refused

To make this work again, switch the order of the entries in the file so that
the ipv4 address comes first.