capi: mask docker service on Flatcar
Docker is preinstalled on Flatcar and the docker.socket unit is enabled by
default. Since kubeadm does probing of available CRI sockets it finds docker
first and enables dockershim instead of using containerd. This was never
intended to work for capi, so explicitly mask the docker units to allow kubeadm
to select containerd instead.