capi: add systemd environment generator to fix-up PATH for Flatcar
The default user PATH on Flatcar Container Linux includes /opt/bin.
image-builder installs a bunch of binaries to /opt/bin, but unfortunately the
default PATH for systemd services is constructed differently and is missing
/opt/bin. To prevent all providers from having to manually append /opt/bin to
service's PATH for every service, add an environment generator script that sets
the PATH to the system-wide default.