capi: ensure kubelet service is stopped
When kubelet is installed on Ubuntu, the service is automatically enabled and
started. But since the configuration is only created by kubeadm when the image
is provisioned, kubelet can't start during the image build and keeps
restarting. The goss test checks that kubelet is not running, and the CI has
been hitting failures because kubelet is running for a short time every 10s.
Ensure the service is stopped to prevent the CI from flaking on us.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>