Link to Windows docs in TOC
mdBook doesn't generate/compile a doc page unless it is referenced from
SUMMARY.md. Even though the windows page was linked to in capi.md, that
wasn't enough to bring it in. By adding it to the Table of Contents, it
shows up in the left-side nav, and the link from capi.md works now too.