[ansible][windows] Add `ExtendVolumesPlugin` cloubase-init plugin
Add `ExtendVolumesPlugin` to the list of default cloudbase-init plugins:
https://cloudbase-init.readthedocs.io/en/latest/plugins.html#cloudbaseinit.plugins.windows.extendvolumes.ExtendVolumesPlugin

The official capi Windows marketplace image from the `cncf-upstream:capi`
are generated without this plugin, and the images have a 30GB system
root partition disk. Using a flavor with a bigger OS disk will
not extend the root system partition, unless manually extended.

The `ExtendVolumesPlugin` plugin is required to extend the root system
to the maximum size by the cloudbase-init.