capi: use generic GOSS validations for Flatcar
It's probably a good idea to use a generic set of validations for the
various providers as much as possible. Flatcar is configured with its
own validations for AWS and Azure at the moment. While working on OVA
support for Flatcar I discovered most generic validations are passing
on Flatcar already. Only the following adaptations had to be made for
Flatcar to use the generic validation rules:

- Flatcar doesn't ship with auditd at the moment.
- Flatcar uses systemd-timesyncd instead of chronyd.

We drop the `crictl ps` validation because there is already a generic
check for that.

Signed-off-by: Johanan Liebermann <jliebermann@microsoft.com>