Update VMware GuestInfo DS to v1.4.1
While this DS has been archived, it did have a v1.4.1 release that
removed the dependency on DeepMerge. This library no longer works
on builds that use Python 2, such as RHEL 7, because the DeepMerge
PyPi releases moved to Python3 packaging, and there is no longer
a setup.py file after release 0.3.0 --
https://pypi.org/project/deepmerge.

This patch just increments the guest info DS to be v1.4.1 so it no
longer attempts to install DeepMerge at all.