images/capi/ansible/roles/setup: override ansible_os_family for Flatcar
Flatcar stable relase since 3139.2.0 has removed
/etc/flatcar/update.conf file from default instalation, which was used
by Ansible to determine the distro type and effectively the value of
'ansible_os_family' fact, which broke building Flatcar images, as now
the conditions for tasks does not evaluate properly when running on
Flatcar, as 'ansible_os_family' is reported as
'Flatcar Container Linux by Kinvolk' instead of just 'Flatcar'.

This commit adds a workaround for this issue by overriding the value for
'ansible_os_family' fact if problematic value is detected.

An alternative to this would be to set a custom fact named e.g.
'is_flatcar' and use that in code, which allows to deduplicate and
encapsulate the logic of determining if Flatcar is in use. However, once
Ansible bug is fixed on determining 'ansible_os_family' value, this code
can be removed and remaining code can stay as is.

See links below for more details:
- https://github.com/flatcar-linux/scripts/pull/212
- https://github.com/ansible/ansible/issues/77537
- https://github.com/kubernetes-sigs/image-builder/issues/862

Refs #862

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>