qemu,raw: Mask sshd during bootstrap
Flatcar ships with sshd enabled by default. Installing Flatcar
entails booting into Flatcar from an ISO and running the installation
command. Packer treats the availability of the SSH port as an
indication that the OS installation is done. To address this, a
static wait time is appended to the Flatcar boot command, which means
there is a potential race between the installation script and Packer.
This leads to occasional Packer errors in Flatcar builds.

Mask sshd during bootstrap. When the installation completes, the VM
is rebooted from disk and then sshd is enabled again automatically.
Also remove the wait as it's no longer necessary because we can rely
on the built-in ssh_timeout Packer parameter like other distros do.

Signed-off-by: Johanan Liebermann <jliebermann@microsoft.com>