Add option for custom pre/post roles
For each playbook (node, windows-node, and firstboot) and an option
pre/post hook for a custom role. This standardizes the way custom roles
are handled across all playbooks, and maintains backwards compatibility
by also making sure that "custom_role_names" is handled just like
"node_custom_roles_post".

The use of the boolean "custom_role" is no longer needed or supported,
as the custom roles are run as needed implicitly bases on whether roles
are defined or empty.