capi,deb,kubevirt: override default netplan for KubeVirt
The default netplan configuration created at cloud-init enables
dhcp for a primary interface, that matches it's mac-address.
However, when used with KubeVirt, the mac address changes on
reboot, which mean that after the reboot, the DHCP is not enabled
and thus IP is not acquired.

This commit overrides the default by enabling the dhcp for ethernet
interfaces that match name for a pattern: enp*s*.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>