Set FLATCAR_VERSION to latest available in Azure CI
Flatcar Azure marketplace publishing frequently lags behind what is
available on Flatcar's servers. In these cases the makefile sets
FLATCAR_VERSION to a value that will fail to provision on Azure because
the image version is not yet available. Override FLATCAR_VERSION to
ensure that the ci is not broken every couple of weeks.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>