Fix typo in artifact_name variable
This corrects a typo in the `artifact_name` user variable, so that it
aligns with the name of `ARTIFACT_NAME` environment variable in the
`build_kubevirt_image.sh` script.

This is in line with the principle of least astonishment in that users
would rightfully expect to be able to configure the `ARTIFACT_NAME`
environment variable by setting `artifact_name` rather than
`artificat_name`.

Signed-off-by: Wolodja Wentland <w@wentland.dev>