Allow users to configure the QEMU VM name
This allows users to configure the name of the VM used by packer
explicitly, which also determines the name of the generated output
file.

Making this configurable is also important in the context of other
settings that have already been exposed to users. The QEMU builds
allow users to configure `artifact_name` which has to be set to the
same value as `vm_name` as it ultimately refers to the aforementioned
output file.

See:
- https://developer.hashicorp.com/packer/plugins/builders/qemu#vm_name

Signed-off-by: Wolodja Wentland <w@wentland.dev>