Disable Flatcar updates during image creation
We want to ensure Flatcar doesn't update to a newer version while
building an image. This can break builds and cause a mismatch between
the version in the AMI metadata and the actual version in the image.

Break down the single Ignition config file we've been using so far into
two files: One for cloud platforms which automatically authorize SSH
keys on created machines and another for platforms which rely on
SSH password authentication.

Use the appropriate Ignition config for each supported platform and
Disable Flatcar updates in both configs.

Signed-off-by: Johanan Liebermann <jliebermann@microsoft.com>