Flatcar: Use OS default volume/exec symlink
The usage of a bind mount does not work with systemd-sysext managing the
/usr mount tree because the mounts may get lost on an extension reload.
Flatcar starts to use extensions by default for cloud integration
software. Fortunately, we don't to define an own systemd-sysext
extension to bring in the symlink under /var because the symlink is
present in all recent Flatcar versions.