Manually execute userdel when deprovisioning Flatcar
Flatcar carries a downstream patch to waagent that was recently reworked and
unfortunately missed testing of the deprovision case. This results in an
exception being thrown. This case can be avoided as long as we manually delete
the user. Remove the '+user' arg and switch to userdel.

This is temporary until we get a fix into Flatcar.