fix: 🐛 Make sure nouveau kernel module is never loaded
By default this operating system auto-loads the nouveau driver when it
detects Nvidia GPU hardware. The nouveau driver sometimes conflicts with
the nvidia proprietary driver that gpu-operator installs. So we need to
blacklist it to prevent from auto-loading during boot and also block
loading when anyother entity tries to load it.