Configure Ubuntu grub defaults like a cloud-image
Images built with image-builder currently use the default grub
configuration which does not enable the serial console, and hides
the boot log with a splash screen on the graphical console.

This patch places grub defaults which match those in the pre-built
ubuntu cloud images, enabling the serial console and disabling the
splash screen.

The boot log is now functional "openstack console log show <server-id>"
and the graphical console is simplified without the splash screen.